briefly/pyproject.toml
[project]
name = "blog-to-pdf"
version = "1.0.0"
description = "Generate 2-page and 3-page PDF briefs from blog posts using FastAPI and ReportLab"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Clancy"}
]

dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",

    # HTTP Client
    "aiohttp>=3.9.0",

    # HTML Parsing
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",

    # PDF Generation
    "reportlab>=4.0.0",

    # Image Processing (for logo brightness detection)
    "Pillow>=10.0.0",

    # LLM Integration
    "litellm>=1.0.0",
    "openai>=1.0.0",

    # Document Processing (brand doc upload)
    "pypdf>=3.17.0",
    "python-docx>=1.0.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/GLN-rc/rcCI-clone/tree/main/blog-pdf"
Repository = "https://github.com/GLN-rc/rcCI-clone"
